+++ src/ZPublisher/tests/testHTTPRequest.py
@@ -808,6 +808,54 @@ class HTTPRequestTests(unittest.TestCase, HTTPRequestFactoryMixin):
         self.assertEqual(req.PATH_INFO, '/test')
         self.assertEqual(req.args, ())
 
+    def test_processInputs_xmlrpc_query_string(self):
+        TEST_METHOD_CALL = (
+            b'<?xml version="1.0"?>'
+            b'<methodCall><methodName>test</methodName></methodCall>'
+        )
+        environ = self._makePostEnviron(body=TEST_METHOD_CALL)
+        environ['CONTENT_TYPE'] = 'text/xml'
+        environ['QUERY_STRING'] = 'x=1'
+        req = self._makeOne(stdin=BytesIO(TEST_METHOD_CALL), environ=environ)
+        req.processInputs()
+        self.assertEqual(req.PATH_INFO, '/test')
+        self.assertEqual(req.args, ())
+        self.assertEqual(req.form["x"], '1')
+
+    def test_processInputs_xmlrpc_method(self):
+        TEST_METHOD_CALL = (
+            b'<?xml version="1.0"?>'
+            b'<methodCall><methodName>test</methodName></methodCall>'
+        )
+        environ = self._makePostEnviron(body=TEST_METHOD_CALL)
+        environ['CONTENT_TYPE'] = 'text/xml'
+        environ['QUERY_STRING'] = ':method=method'
+        req = self._makeOne(stdin=BytesIO(TEST_METHOD_CALL), environ=environ)
+        with self.assertRaises(BadRequest):
+            req.processInputs()
+
+    def test_processInputs_SOAP(self):
+        # ZPublisher does not really have SOAP support
+        # all it does is put the body into ``SOAPXML``
+        body = b'soap'
+        environ = TEST_POST_ENVIRON.copy()
+        environ['HTTP_SOAPACTION'] = "soapaction"
+        req = self._makeOne(stdin=BytesIO(body), environ=environ)
+        req.processInputs()
+        self.assertEqual(req.SOAPXML, body)
+
+    def test_processInputs_SOAP_query_string(self):
+        # ZPublisher does not really have SOAP support
+        # all it does is put the body into ``SOAPXML``
+        body = b'soap'
+        environ = TEST_POST_ENVIRON.copy()
+        environ['QUERY_STRING'] = 'x=1'
+        environ['HTTP_SOAPACTION'] = "soapaction"
+        req = self._makeOne(stdin=BytesIO(body), environ=environ)
+        req.processInputs()
+        self.assertEqual(req.SOAPXML, body)
+        self.assertEqual(req.form["x"], '1')
+
     def test_processInputs_w_urlencoded_and_qs(self):
         body = b'foo=1'
         environ = {
@@ -858,6 +906,46 @@ class HTTPRequestTests(unittest.TestCase, HTTPRequestFactoryMixin):
         self.assertEqual(list(f), [b'test\n'])
         f.seek(0)
         self.assertEqual(next(f), b'test\n')
+
+    def test_processInputs_BODY(self):
+        s = BytesIO(b"body")
+        environ = TEST_POST_ENVIRON.copy()
+        environ["CONTENT_TYPE"] = "text/plain"
+        req = self._makeOne(stdin=s, environ=environ)
+        req.processInputs()
+        self.assertEqual(req["BODY"], b"body")
+        self.assertIs(req["BODYFILE"], s)
+
+    def test_processInputs_BODY_unseekable(self):
+        s = _Unseekable(BytesIO(b"body"))
+        environ = TEST_POST_ENVIRON.copy()
+        environ["CONTENT_TYPE"] = "text/plain"
+        req = self._makeOne(stdin=s, environ=environ)
+        req.processInputs()
+        self.assertEqual(req["BODY"], b"body")
+        self.assertIs(req["BODYFILE"], s)
+
+    def test_processInputs_seekable_form_data(self):
+        s = BytesIO(TEST_FILE_DATA)
+        environ = self._makePostEnviron(body=TEST_FILE_DATA)
+        req = self._makeOne(stdin=s, environ=environ)
+        req.processInputs()
+        f = req.form.get('smallfile')
+        self.assertEqual(list(f), [b'test\n'])
+        self.assertEqual(req["BODY"], TEST_FILE_DATA)
+        self.assertEqual(req["BODYFILE"].read(), TEST_FILE_DATA)
+
+    def test_processInputs_unseekable_form_data(self):
+        s = _Unseekable(BytesIO(TEST_FILE_DATA))
+        environ = self._makePostEnviron(body=TEST_FILE_DATA)
+        req = self._makeOne(stdin=s, environ=environ)
+        req.processInputs()
+        f = req.form.get('smallfile')
+        self.assertEqual(list(f), [b'test\n'])
+        # we cannot access ``BODY`` in this case
+        # as the underlying file has been read
+        with self.assertRaises(KeyError):
+            req["BODY"]
 
     def test__authUserPW_simple(self):
         user_id = 'user'
@@ -1347,7 +1435,7 @@ class HTTPRequestTests(unittest.TestCase, HTTPRequestFactoryMixin):
         req.processInputs()
         self.assertDictEqual(req.form, {"bar": ""})
 
-    def test_put_with_body_and_query_string_raises(self):
+    def test_put_with_body_and_query_string(self):
         req_factory = self._getTargetClass()
         req = req_factory(
             BytesIO(b"foo"),
@@ -1355,12 +1443,13 @@ class HTTPRequestTests(unittest.TestCase, HTTPRequestFactoryMixin):
                 "SERVER_NAME": "localhost",
                 "SERVER_PORT": "8080",
                 "REQUEST_METHOD": "PUT",
-                "QUERY_STRING": "bar"
+                "QUERY_STRING": "bar=bar"
             },
             None,
         )
-        with self.assertRaises(NotImplementedError):
-            req.processInputs()
+        req.processInputs()
+        self.assertEqual(req.BODY, b"foo")
+        self.assertEqual(req.form["bar"], "bar")
 
     def test_issue_1095(self):
         body = TEST_ISSUE_1095_DATA
@@ -1449,6 +1538,13 @@ class TestSearchType(unittest.TestCase):
 
     def test_special(self):
         self.check("abc:a-_0b", ":a-_0b")
+
+
+class _Unseekable:
+    """Auxiliary class emulating an unseekable file like object"""
+    def __init__(self, file):
+        for m in ("read", "readline", "close", "__del__"):
+            setattr(self, m, getattr(file, m))
 
 
 TEST_POST_ENVIRON = {